pyproject.toml
[project]
name = "cxview-core"
version = "0.1.0"
description = "Shared core library for microservices"
authors = [
  { name = "Ailen Vo", email = "ailenvo@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
  "pydantic>=2.0",
  "uuid6>=2025.0.1",
  "pytz>=2024.2",
  "python-jose>=3.3.0",
  "motor>=3.6.0",
  "pymongo>=4.9.2",
  "pydantic>=2.9.2",
  "SQLAlchemy>=2.0.44",
]
